core/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.14",
    "fastapi>=0.115.12",
    "openai>=1.69.0",
    "pydantic-settings>=2.8.1",
    "pymysql>=1.1.1",
    "snowflake-id>=1.0.2",
    "sqlalchemy==2.0.30",
    "redis-py-cluster>=2.1.3",
    "uvicorn>=0.34.0",
    "loguru>=0.7.2",
    "sqlmodel>=0.0.14",
    "confluent-kafka>=2.5.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
    "opentelemetry-proto>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-semantic-conventions>=0.46b0",
    "opentelemetry-exporter-opencensus>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "grpc-google-iam-v1==0.12.6",
    "googleapis-common-protos==1.60.0",
    "websocket-client==1.8.0",
    "python-dotenv>=1.0.1",
    "toml>=0.10.2",
    "websockets==12.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.4.0",
    "httpx>=0.27.0",
    "faker>=26.0.0",
    "boto3>=1.40.53",
    "flake8==7.0.0",
    "isort==5.13.2",
    "black==24.4.2",
    "mypy==1.18.2",
    "pylint==3.1.0",
]
